refactor: Moved field styling logic to a separate file
